requirements.txt
cx_Oracle
django-crispy-forms